Spectrograms

A basic function is provided to plot spectrograms that look familiar to phoneticians. It makes use of the spectrogram function from DSP.jl to perform the short-time Fourier analysis.

Examples

A standard broadband spectrogram can be created without using optional parameters.

using Phonetics # hide
using WAV
s, fs = wavread("assets/iwantaspectrogram.wav")
s = vec(s)
phonspec(s, fs)

A color scheme more similar to the Praat grayscale can be achieved using the col argument:

using Phonetics # hide
using WAV # hide
s, fs = wavread("assets/iwantaspectrogram.wav") # hide
s = vec(s) # hide
using Plots
rev_grays = reverse(cgrad(:grays))
phonspec(s, fs, col=rev_grays)

A narrowband style spectrogram can be plotted using the style argument:

using Phonetics # hide
using WAV # hide
s, fs = wavread("assets/iwantaspectrogram.wav") # hide
s = vec(s) # hide
phonspec(s, fs, style=:narrowband)

And, the pre-emphasis can be disabled by passing in a value of 0 for the pre_emph argument:

using Phonetics # hide
using WAV # hide
s, fs = wavread("assets/iwantaspectrogram.wav") # hide
s = vec(s) # hide
phonspec(s, fs, pre_emph=0)

Function documentation

Phonetics.phonspec — Method

phonspec(s, fs; pre_emph=0.97, col=:magma, style=:broadband, dbr=55, size=(600, 400))

Rudimentary functionality to plot a spectrogram, with parameters familiar to phoneticians. Includes a pre-emphasis routine which helps increase the intensity of the higher frequencies in the display. Uses a Kaiser window with a parameter value of 2.

Args

s A vector containing the samples of a sound

fs Sampling frequency of s in Hz

pre_emph The α coefficient for pre-emmphasis; default value of 0.97 corresponds to a cutoff frequency of approximately 213 Hz before the 6 dB / octave increase begins

col Color scheme for the spectrogram

style Either :broadband or :narrowband; will affect the window length and window stride

dbr The dynamic range; all frequencies that are dbr decibels quieter than the loudest frequency will not be displayed

size Size of plot in pixels; passed to heatmap call